move test helpers to helpers_test.go
